pyproject.toml
[tool.poetry]
name = "sleep-tracking"
version = "0.1.0"
description = "Sleep stage estimation using RNNs with physiological data from wearable devices"
authors = ["pashupati.hegde@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
matplotlib = "^3.4.3"
numpy = "^1.21.4"
pandas = "^1.3.4"
scikit-learn = "^1.0.1"
scipy = "^1.7.2"
seaborn = "0.11.1"
statsmodels= "^0.13.1"
torch = "1.9.0"
tqdm = "^4.56.0"
pdfminer = "^20191125"
docx2txt = "^0.8"
pytorch-ignite = "^0.4.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
